requirements.txt

# general modules
Django==4.2
python-decouple
pillow
djangorestframework==3.14


# third party modules
markdown
django-filter  
drf-yasg==1.21.8


# deployment modules
